Add files via upload
This is the current zip file that I have for the website I was working on the home page for a while then started working on the log in/ sign up pages but I am running into the problem that my files aren't connecting to the local server I will try to work on it on my lunch break at work hopefully it will get fixed. I did download the program XAMPP to be able to connect to the local server... http://localhost/phpmyadmin/ 
I am adding a screenshot of the server with the current table for the user hopefully I can get it connected to the server. Let me know if yall need the sql file from the server..