Add files via upload
Im sending a zip again because all of the pages had some update little or big, I also added the powerpoint for tomorrow im sure you guys can edit or add more to it if needed. I couldnt test the database so I hope the reservations page is good, as well as the users page, that is if they log in.